+++ src/lib_injector/injector_sigs.ml
@@ -369,9 +369,17 @@ module type S = sig
       tag. *)
   val get_queues : ?tag:tag -> unit -> (tag list * Inj_operation.t list) list
 
-  (** Clears the injectors queues completely. If [tag] is provided, only queues
-      for the injector which handles this tag is cleared. *)
-  val clear_queues : ?tag:tag -> unit -> unit tzresult Lwt.t
+  (** Clears the injectors queues completely. If [tag] is provided,
+      only queues for the injector which handles this tag is
+      cleared. If [order_below], only operation with order below that
+      value are cleared, and if [drop_no_order] is true, clear all
+      operation that has no order specified. *)
+  val clear_queues :
+    ?drop_no_order:bool ->
+    ?order_below:Z.t ->
+    ?tag:tag ->
+    unit ->
+    unit tzresult Lwt.t
 
   (** Register a protocol client for a specific protocol to be used by the
       injector. This function {b must} be called for all protocols that the